requirements.txt
altgraph==0.17.4
asttokens==2.4.1
attrs==24.2.0
auto-py-to-exe==2.44.1
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
bottle==0.13.1
bottle-websocket==0.2.9
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
cryptography==43.0.1
debugpy==1.8.5
defusedxml==0.7.1
distlib==0.3.8
docopt==0.6.2
Eel==0.17.0
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.14.0
fonttools==4.54.0
fpdf2==2.7.9
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pefile==2024.8.26
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==5.29.0
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
rsa==4.9
selenium==4.25.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.2
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.44.0
whichcraft==0.6.1
wsproto==1.2.0
yarg==0.1.9
zope.event==5.0
zope.interface==7.0.3
